Moving TargetEncoder to the package ai.h2o.sparkling.ml
