+++ ml/src/main/scala/ai/h2o/sparkling/ml/features/H2OTargetEncoderBase.scala
@@ -1,25 +1,8 @@
-/*
-* Licensed to the Apache Software Foundation (ASF) under one or more
-* contributor license agreements.  See the NOTICE file distributed with
-* this work for additional information regarding copyright ownership.
-* The ASF licenses this file to You under the Apache License, Version 2.0
-* (the "License"); you may not use this file except in compliance with
-* the License.  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
+package ai.h2o.sparkling.ml.features
 
-package org.apache.spark.ml.h2o.features
-
+import ai.h2o.sparkling.ml.params.H2OTargetEncoderParams
 import org.apache.spark.h2o.Frame
 import org.apache.spark.ml.PipelineStage
-import org.apache.spark.ml.h2o.param.H2OTargetEncoderParams
 import org.apache.spark.sql.types.{DoubleType, StructField, StructType}
 
 trait H2OTargetEncoderBase extends PipelineStage with H2OTargetEncoderParams {
